+++ source/renderer/app/containers/wallet/dialogs/WalletImportDialogContainer.js
@@ -1,18 +1,32 @@
 // @flow
 import React, { Component } from 'react';
 import { observer, inject } from 'mobx-react';
+import { map } from 'lodash';
 import WalletImportFileDialog from '../../../components/wallet/wallet-import/WalletImportFileDialog';
 import WalletSelectImportDialog from '../../../components/wallet/wallet-import/WalletSelectImportDialog';
 import { isValidWalletName } from '../../../utils/validations';
 import type { InjectedProps } from '../../../types/injectedPropsType';
 import type { ImportFromOption } from '../../../types/walletExportTypes';
+import { MAX_ADA_WALLETS_COUNT } from '../../../config/numbersConfig';
+import { WalletImportStatuses } from '../../../types/walletExportTypes';
 
 type Props = InjectedProps;
 
+type State = {
+  existingWalletsCount: number,
+};
+
 @inject('stores', 'actions')
 @observer
-export default class WalletImportDialogContainer extends Component<Props> {
+export default class WalletImportDialogContainer extends Component<
+  Props,
+  State
+> {
   static defaultProps = { actions: null, stores: null };
+
+  state = {
+    existingWalletsCount: this.props.stores.wallets.all.length,
+  };
 
   componentWillUnmount() {
     this.props.actions.walletMigration.finishMigration.trigger();
@@ -61,6 +75,18 @@ export default class WalletImportDialogContainer extends Component<Props> {
       walletMigrationStep,
     } = walletMigration;
     const { openExternalLink } = app;
+    let walletsCount =
+      this.state.existingWalletsCount + pendingImportWalletsCount;
+    map(exportedWallets, wallet => {
+      if (
+        wallet.import.status === WalletImportStatuses.COMPLETED ||
+        wallet.import.status === WalletImportStatuses.RUNNING
+      ) {
+        walletsCount++;
+      }
+    });
+    const isMaxNumberOfWalletsReached =
+      walletsCount >= MAX_ADA_WALLETS_COUNT;
 
     return (
       <>
@@ -90,6 +116,7 @@ export default class WalletImportDialogContainer extends Component<Props> {
             onWalletNameChange={this.onWalletNameChange}
             onToggleWalletImportSelection={this.onToggleWalletImportSelection}
             onClose={this.onCancel}
+            isMaxNumberOfWalletsReached={isMaxNumberOfWalletsReached}
           />
         )}
       </>