+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog
 
 ### Chores
 
+- Prevented import for more than maximum number of wallets ([PR 2012](https://github.com/input-output-hk/daedalus/pull/2012))
 - Implemented automated tests for wallet migration and import features ([PR 2011](https://github.com/input-output-hk/daedalus/pull/2011))
 
 ## 1.1.0-FC1